fixed deepMergind
