+++ packages/helpers/src/keyCodes/keyCodes.ts
@@ -1,0 +1,21 @@
+/**
+ * Returns basic key codes as constants
+ * @returns {{}}
+ */
+export const keyCodes = {
+  BACKSPACE: 8,
+  TAB: 9,
+  ENTER: 13,
+  SHIFT: 16,
+  CTRL: 17,
+  ALT: 18,
+  ESC: 27,
+  SPACE: 32,
+  LEFT: 37,
+  UP: 38,
+  DOWN: 40,
+  RIGHT: 39,
+  DELETE: 46,
+  META: 91,
+  SLASH: 191,
+};
