+++ packages/helpers/src/keyCodes/mouseButtons.ts
@@ -1,0 +1,10 @@
+/**
+ * Return mouse buttons codes
+ */
+export const mouseButtons = {
+  LEFT: 0,
+  WHEEL: 1,
+  RIGHT: 2,
+  BACKWARD: 3,
+  FORWARD: 4,
+};
